backend/requirements.txt
# Resolwe server requirements (compatible with Python 3.11)
resolwe>=43.0.0,<44.0.0
resolwe-bio
djangorestframework>=3.14.0
django-filter>=23.0.0
django-cors-headers>=4.0.0
celery>=5.3.0
redis>=5.0.0
psycopg[binary]>=3.1.0
channels>=4.0.0
channels-redis>=4.0.0
python-decouple>=3.8
daphne>=4.0.0
django-guardian>=2.4.0
